% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert_gt}
\alias{convert_gt}
\title{Convert genotype to numeric values between 0.0 and 2.0}
\usage{
convert_gt(gt, type)
}
\arguments{
\item{gt}{when it's character, this function converts to numerical values
representation according to type FAILED, FALSE, and other types will be
converted to NA; when it's numerics, then conversion is directly peformed}

\item{type}{must be one of the three: "additive", "recessive" and "dominant"}
}
\description{
Convert genotype to numeric values between 0.0 and 2.0
}
\keyword{internal}

