% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{convert_tail}
\alias{convert_tail}
\title{this function converts a two-tailed p value to a one-tailed p value}
\usage{
convert_tail(direction, pval, tail)
}
\arguments{
\item{direction}{is positive}

\item{pval}{pvalue}

\item{tail}{"decreasing", "increasing" or "both"}
}
\description{
this function converts a two-tailed p value to a one-tailed p value
}

