% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_data.R
\name{dichotomize}
\alias{dichotomize}
\title{Dichotomize a numeric vector into AFFECTED/UNAFFECTED}
\usage{
dichotomize(x, log_file = NULL)
}
\arguments{
\item{x}{numeric vector}

\item{log_file}{specify log file}
}
\value{
list(), group -> suggested group number
}
\description{
Dichotomize a numeric vector into AFFECTED/UNAFFECTED
}

