% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fit.model.R
\name{fit.null.model}
\alias{fit.null.model}
\title{if success, will return append "result" field to}
\usage{
fit.null.model(null.model, pheno)
}
\arguments{
\item{null.model}{otherwise, will return a list including error informatino}
}
\description{
if success, will return append "result" field to
}

