% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gene.single.link}
\alias{gene.single.link}
\title{Meta-analyze single gene in super pedigree}
\usage{
gene.single.link(vcfFile, pedFile, pheno.name, output = ".", test = "wG2",
  detect = "never", silent = T, tail = "decreasing", prefix = "",
  plot.it = TRUE, transform.pheno = NULL, log.level = "WARN")
}
\arguments{
\item{vcfFile}{genotype input file in VCF format}

\item{pedFile}{pedigree file in PED format}

\item{pheno.name}{phenotype name to analyze}

\item{output}{The output folder to put all the output files of the analysis}

\item{test}{The statistical test to be used for identifying significant genes
associated with phenotype. Default is "wG2", considering G2 mother
effect. Another option is "woG2", not considering that effect.}

\item{detect}{A character string specifying whether to detect clustering of
phenotypic scores and to transform continuous phenotypic scores into a binary
variable (affected and nonaffected). This parameter only works on continuous
phenotype scores. "never" (default): never transform; "always": always
transform; "auto": let the program decide whether to transform.}

\item{silent}{Print intermediate messages to stdout if set to TRUE.}

\item{tail}{Either "decreasing", "increasing" or "both". "decreasing" tests
whether the mutation leads to decreased antibody reaction (default);
"increasing" tests whether the mutation leads to increased antibody reaction;
"both" tests the deviation in either direction.}

\item{prefix}{Default is "". An optional character string to be attached to
output file names. This can create personalized names for each job.}

\item{plot.it}{Default is TRUE, it controls whether to output linkage plots
and distribution plots.}

\item{transform.pheno}{Default is NULL. Use "log" if phenotypes need to log
transformed.}

\item{log.level}{Default WARN, but can be set from 'DEBUG', 'INFO', 'WARN'
and 'ERROR'.}
}
\value{
a list of three values will be returned:
the first one is returncode (0: success, non-zero: errors happened)
the second one is message, when return is non-zero, message gives explanations
the last one is analysis results.
}
\description{
Perform gene-based superpedigree analysis
}
\examples{
path <- system.file("extdata/gene",package="LinkageAnalyzer")
vcfFile <- file.path(path, "Snrnp40.mcmv.vcf")
pedFile <- file.path(path, "Snrnp40.mcmv.ped")
pheno.name <- "log"
output <- file.path(path, "output")
ret <- gene.single.link(vcfFile, pedFile, pheno.name, output,
                        test = "woG2", tail = "both", prefix="gene")
}

