% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_data.R
\name{geno.from.012}
\alias{geno.from.012}
\title{Convert 0,1,2,NA to "REF", "HET", "ALT", "FAIL"}
\usage{
geno.from.012(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
character vector
}
\description{
Convert 0,1,2,NA to "REF", "HET", "ALT", "FAIL"
}
\keyword{internal}

