% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_data.R
\name{get.ped}
\alias{get.ped}
\title{read a PED file}
\usage{
get.ped(pedFile, pheno = NULL, detect = NULL)
}
\arguments{
\item{pedFile}{file name in PED format}

\item{pheno}{phenotype column header}

\item{detect}{"auto" means detect binary phenotype, or "NULL" means do nothing}
}
\value{
a data frame that looks like PED file content
}
\description{
read a PED file
}

