% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_G2}
\alias{get_G2}
\title{function to read G2 mice genotypes}
\usage{
get_G2(file = "", log_file)
}
\arguments{
\item{file}{a CSV file listing G2 mothers' genotypes}

\item{log_file}{specify log file}
}
\value{
a list of
returncode: non-zero when criitical error
message: related to the function running status
data: data loaded
}
\description{
function to read G2 mice genotypes
}
\keyword{internal}

