% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_data.R
\name{order.chrom.pos}
\alias{order.chrom.pos}
\title{Order by chromosomal positions}
\usage{
order.chrom.pos(chrom, pos)
}
\arguments{
\item{chrom}{chacter vector listing chromosome names, e.g. "1", "2", ...}

\item{pos}{integer vector}
}
\value{
indice which can be used to sort
}
\description{
Order by chromosomal positions
}

