% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mht_plot.R
\name{plot.manhattan}
\alias{plot.manhattan}
\title{Draw manhattan plot}
\usage{
\method{plot}{manhattan}(data, main = "")
}
\arguments{
\item{data}{a data frame with at least three columns: Chrom, Position, Pval}

\item{main}{plot title}
}
\description{
Draw manhattan plot
}

