% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{reloadLibrary}
\alias{reloadLibrary}
\title{Quick reload library in R}
\usage{
reloadLibrary(name, dynam = FALSE)
}
\arguments{
\item{dynam}{whether also unload dynamic library}
}
\description{
Quick reload library in R
}

