% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/single_link.R
\name{single.link.impl}
\alias{single.link.impl}
\title{Implementation of single.link()}
\usage{
single.link.impl(vcfFile, pedFile, pheno.name, output = ".", test = "wG2",
  detect = "never", silent = T, tail = "decreasing", prefix = "",
  plot.it = TRUE, transform.pheno = NULL, log.level = "WARN")
}
\description{
Implementation of single.link()
}

