% Generated by roxygen2 (4.0.2): do not edit by hand
\name{single.link}
\alias{single.link}
\title{Statistical analysis for correlation with phenotype and lethality on a single locus}
\usage{
single.link(vcfFile, pedFile, pheno.name, output = ".", test = "wG2",
  detect = "never", silent = T, tail = "decreasing", prefix = "",
  plot.it = TRUE, transform.pheno = NULL, log.level = "WARN")
}
\arguments{
\item{vcfFile}{genotype input file in VCF format}

\item{pedFile}{pedigree file in PED format}

\item{pheno.name}{phenotype name to analyze}

\item{output}{The output folder to put all the output files of the analysis}

\item{test}{The statistical test to be used for identifying significant genes
associated with phenotype. Default is "wG2", considering G2 mother effect.
Another option is "woG2", not considering that effect.}

\item{detect}{A character string specifying whether to detect clustering of
phenotypic scores and to transform continuous phenotypic scores into a binary
variable (affected and nonaffected). This parameter only works on continuous
phenotype scores. "never" (default): never transform; "always": always
transform; "auto": let the program decide whether to transform.}

\item{silent}{Print intermediate messages to stdout if set to TRUE.}

\item{tail}{Either "decreasing", "increasing" or "both". "decreasing" tests
whether the mutation leads to decreased antibody reaction (default);
"increasing" tests whether the mutation leads to increased antibody reaction;
"both" tests the deviation in either direction.}

\item{prefix}{Default is "". An optional character string to be attached to
output file names. This can create personalized names for each job.}

\item{plot.it}{Default is TRUE, it controls whether to output linkage plots
and distribution plots.}

\item{transform.pheno}{Default is NULL. Use "log" if phenotypes need to log
transformed.}

\item{log.level}{Default WARN, but can be set from 'DEBUG', 'INFO', 'WARN'
and 'ERROR'.}
}
\value{
a list of two values will be returned. One is returncode (0: success).
The other is analysis results.
}
\description{
This function reads the raw data and conducts statistical
analysis for the significance of correlation with phenotype and homozygous
lethality for each gene.
}
\details{
single.link takes a VCF file as genotype covariates and a PED format
as phenotype (response).  A sample input file is stored in the
\code{inst/extdata/single} folder of this package. The algorithm will
calculate the p values for lethality conditioning on G2 dam mice genotype
for a certain gene. The final p value is the combined p values from all the
G3 mice.  A PDF file will be generated. It will contain a diagnostic plot if
continuous phenotype scores are provided. It will then contain three
manhattan plots, each assuming one of the three "additive", "recessive" and
"dominant" models. The two cutoff values are 0.05 and 0.05/(number of genes).
The p values are transformed to a -log10 scale for better visualization.
It also contain a plot for the analysis results from TDT analysis.
A CSV file containing will be generated. Columns names are in the first
line. The REF, HET, VAR columns are the number of mice with each genotype for
each gene. The "lethal" column is the p value for testing homozygous
lethality. The "additive", "recessive" and "dominant" columns are the p
values for testing the significance of the correlation of each gene and the
phenotype under different models. The "TDT" column is the p value of TDT test
for testing each gene's significance. The "Penetrance_REF", "Penetrance_HET",
"Penetrance_VAR" columns are the proportions of G3 mice with each genotype
showing the "affected" phenotype (only applies to binary phenotype scores or
binary scores that are converted from continuous phenotype scores). The
"Semidominance" column is the semidominance of the mutations, only calculated
when continuous phenotype scores are
provided. Semidominance=(HET-REF)/(VAR-REF). 0 means the HET phenotype is the
same as or over the average of the REF phenotype and 1 means the HET
phenotype is the same as or over the average of the VAR phenotype. Note that
VAR may not necessarily be lower than REF.  A second PDF file will be
generated containing scattorplots of phenotypes scores (continuous variable)
or table plots of phenotypes (binary variable).  A TXT file containing
messages, warnings and errors of the program will be generated.  The
statistical model used to calculate the p value of lethality, more accurately
speaking, is testing "whether homozygous mutation will have any detrimental
effect on the survival of mice at the time of data collection". chrX genes
are ignored for the moment.
}
\examples{
path <- system.file("extdata/single",package="LinkageAnalyzer")
vcfFile <- file.path(path, "R0491_body_weight.vcf")
pedFile <- file.path(path, "R0491_body_weight.ped")
pheno.name <- "weight"
output <- file.path(path, "output")
ret <- single.link(vcfFile, pedFile, pheno.name, output,
                   test = "woG2", tail = "both", prefix="single")
}

