% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/single_parse.R
\name{single_parse}
\alias{single_parse}
\title{Parse the analysis results of multiple runs of single.link}
\usage{
single_parse(dir = ".", runs = c(), output = ".", silent = FALSE,
  prefix = "")
}
\arguments{
\item{dir}{The main directory where there are at least one sub-folder which
contain(s) the analysis results of \code{single.link}}

\item{runs}{A vector of character strings. The sub-folder name(s) directly
under \code{dir} containing the analysis results of \code{single.link} that
should be pooled and parsed. WARNING: under each folder specified by
\code{runs}, there can only be the output files of one run of
\code{single.link}!}

\item{output}{The output folder to put the summarized output}

\item{silent}{Print intermediate messages to stdout if set to FALSE.}

\item{prefix}{Default is "". An optional character string to be attached to
output file names. This can create personalized names for each job.}
}
\value{
No return value.
}
\description{
This function loads the RData file under each folder for one
\code{single.link} analysis event and make a summary of the analysis results
for multiple runs. WARNING: under each folder specified by \code{runs}, there
can only be the output files of one run of \code{single.link}!
}
\details{
The analysis output from multiple runs of \code{single.link} will be
parsed and summarized into a .csv file to be put under \code{output}. A .txt
log file will also be generated.
}
\examples{
\dontrun{
 dir <- system.file("extdata/single/output",package="LinkageAnalyzer")
 single_parse(dir=dir,runs=c("raw_continuous_woG2",
                             "raw_continuous_wG2",
                             "autoBinary_norm_continuous_woG2"),output=dir)
}
}
\seealso{
\code{\link{single.link}}
}
\keyword{internal}

