% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{source.all.file}
\alias{source.all.file}
\title{re-source all .R files}
\usage{
source.all.file(dir = ".")
}
\arguments{
\item{dir}{directory name}
}
\description{
re-source all .R files
}
\keyword{internal}

