% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_data.R
\name{vcf.add.sample}
\alias{vcf.add.sample}
\title{Add samples to vcf data and set their genotypes as missing (GT, REF, VAR)}
\usage{
vcf.add.sample(vcf, sampleName)
}
\arguments{
\item{vcf}{vcf data}

\item{sampleName}{character vector}
}
\value{
vcf data
}
\description{
Add samples to vcf data and set their genotypes as missing (GT, REF, VAR)
}

