% Generated by roxygen2 (4.0.2): do not edit by hand
\name{crossCheckMotherGenotype}
\alias{crossCheckMotherGenotype}
\title{Examine mother-offspring pair and fix mother's genotype by her offsprings.
e.g. mom has VAR offsprings, mom genotype cannot be REF => mom must be HET}
\usage{
crossCheckMotherGenotype(pheno, geno.name = "gt")
}
\arguments{
\item{pheno}{ped data}

\item{geno.name}{specify genotype column headers}
}
\description{
Examine mother-offspring pair and fix mother's genotype by her offsprings.
e.g. mom has VAR offsprings, mom genotype cannot be REF => mom must be HET
}
\keyword{internal}

